Move recipe loading into own build manager and create runner to handle invoking builds
